test fix

+++ cpm_lib/include/cpm/MultiVehicleReader.hpp
@@ -63,7 +63,7 @@ namespace cpm
             }
         }
 
-        MultiVehicleReader(dds::topic::Topic<T> topic, std::vector<int> _vehicle_ids) : 
+        MultiVehicleReader(dds::topic::Topic<T> topic, std::vector<uint8_t> _vehicle_ids) : 
             dds_reader(dds::sub::Subscriber(ParticipantSingleton::Instance()), topic, (dds::sub::qos::DataReaderQos() << dds::core::policy::History::KeepAll())
         )
         {
